backend/requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
ultralytics>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0
requests==2.31.0

# LiveKit dependencies
livekit-api>=0.5.0
livekit>=0.10.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-silero>=0.6.0
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0